Potential refactor to look out styled components
